requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
anyio==3.7.1
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.104.1
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.25.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.3
h11==0.14.0
h5py==3.10.0
idna==3.6
importlib-metadata==7.0.0
jax==0.4.13
keras==2.12.0
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
ml-dtypes==0.2.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
Pillow==10.1.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
python-multipart==0.0.6
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0
